requirements.txt
python-dotenv==0.20.0
discord.py==1.7.3
httpx==0.23.0
